requirements.txt
matplotlib                   
numpy                        
pandas                       
pycparser                   
pyparsing                    
PyYAML                       
python-dotenv                
scikit-learn                  
torch                        
torchaudio                    
torchinfo                     
torchvision                   
timm